Rename CLI entry to src/cli/run.ts.
Restore evals task path and update imports and README layout.
