+++ src/cli/run.ts
