+++ tests/cli/options.test.ts
@@ -1,9 +1,6 @@
 import { assertEquals, assertThrows } from "@std/assert";
 import { evalCases } from "../../src/cases/index.ts";
-import {
-  parseCliOptions,
-  selectEvalCases,
-} from "../../src/cli/parse-cli-options.ts";
+import { parseCliOptions, selectEvalCases } from "../../src/cli/run.ts";
 
 Deno.test("parseCliOptions rejects removed golden flags", () => {
   assertThrows(
